refactor: extract TerminalRenderer from TerminalPlugin
Move all rendering logic (ANSI helpers, markdown conversion, code block
rendering, DSL rendering, block rendering) from TerminalPlugin into a new
TerminalRenderer struct. TerminalPlugin now delegates to the renderer.

Source: user
Type: refactor
