+++ src/cli/mod.rs
@@ -4,6 +4,7 @@ pub mod admin;
 pub mod args;
 pub mod chat;
 pub mod config_wizard;
+pub mod renderer;
 pub mod terminal;
 
 #[cfg(test)]
